requirements.txt
openai==0.27.4
requests==2.28.2
discord.py==2.1.1
python-dotenv==0.21.1
Flask==2.2.3
